refactor: Sections 일급 컬렉션으로 로직 이동
